+++ types/merchandise.d.ts
@@ -1,9 +1,15 @@
 import { APIOptions } from "yonius";
 import { Base, BaseDelta } from "./base";
 
-export class Merchandise extends Base {}
+export class Merchandise extends Base {
+    company_product_code: string;
+    barcode: string;
+}
 
-export class MerchandiseDelta extends BaseDelta {}
+export class MerchandiseDelta extends BaseDelta {
+    company_product_code?: string;
+    barcode?: string;
+}
 
 export class MerchandisePayload {
     merchandise?: MerchandiseDelta;